Every possible combination of rotational states for each object is now computed.
Additionally the program filters out duplicate rotations (rotations that achieve the exact same object as a previous rotation) and rotations that result in an object that is guaranteed to not fit on the playing field. This cut possible unique cases down from the theoretical maximum of 64^4 (almost 17 million) to only 49 152 with the current test blocks (this will slightly vary with different blocks).